+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.1
+
+- fix: one endpoint with different methods
+
+# 2.0.0
+
+- simplified router
+
 # 1.0.2
 
 - fix: reload